+++ Sources/TetrisEngine/Configuration.swift
@@ -1,0 +1,13 @@
+//
+// Configuration.swift
+// Created by: Jérémy Goyette
+// On: 2021/02/22
+// For: Tetris Engine
+//
+// G-Labs
+//
+
+import CTetrisEngine
+
+/// The settings used to create a Tetris game.
+public typealias Configuration = TetrisGameConfiguration